+++ projects/scalar-app/src/features/app/hooks/use-version-conflict-check.test.ts
@@ -3,9 +3,9 @@ import { flushPromises } from '@vue/test-utils'
 import { describe, expect, it, vi } from 'vitest'
 import { nextTick, ref } from 'vue'
 
-import type { ImportDocumentFromRegistry } from '@/types/configuration'
 import type { SidebarDocumentVersion } from '@/features/app/hooks/use-sidebar-documents'
 import { useVersionConflictCheck } from '@/features/app/hooks/use-version-conflict-check'
+import type { ImportDocumentFromRegistry } from '@/types/configuration'
 
 type FakeDocument = {
   'x-scalar-registry-meta'?: {
