+++ projects/scalar-app/src/features/app/hooks/use-version-conflict-check.ts
@@ -1,9 +1,9 @@
 import type { WorkspaceStore } from '@scalar/workspace-store/client'
 import { type MaybeRefOrGetter, toValue, watch } from 'vue'
 
-import type { ImportDocumentFromRegistry } from '@/types/configuration'
 import { checkVersionConflict } from '@/features/app/helpers/check-version-conflict'
 import type { SidebarDocumentVersion } from '@/features/app/hooks/use-sidebar-documents'
+import type { ImportDocumentFromRegistry } from '@/types/configuration'
 
 /**
  * Kicks off background `checkVersionConflict` calls for every loaded
